django/requirements.txt
django
requests
urllib3
gunicorn
psycopg2-binary
django-environ
markdown
django-filter
pytz
djangorestframework
